update test casts

+++ test/libsolidity/smtCheckerTests/functions/function_external_call_should_not_inline_1.sol
@@ -9,7 +9,7 @@ contract State {
 contract C {
 	address owner;
 	function setOwner() public {
-		owner = address(0);
+		owner = address(uint160(0));
 	}
 }
 // ----
